No need to use long() anymore now that we use cmp() instead of '-'. Excellent.
Revert "Use long() instead of int() for version_int, so we do not exceed sys.maxint"

This reverts commit 8e65a4ddeb74611f7707268a3825a2c6f972299c.

+++ apps/feedback/version_compare.py
@@ -115,7 +115,7 @@ def version_int(version):
     v = "%d%02d%02d%02d%d%02d%d%02d" % (d['major'], d['minor1'],
             d['minor2'], d['minor3'], d['alpha'], d['alpha_ver'], d['pre'],
             d['pre_ver'])
-    return long(v)
+    return int(v)
 _version_int_cache = {}
 version_int = memoize(version_int, _version_int_cache, 1)
